Facing issues while UPDATING and DELETING THE Contacts - Fix/Resolution required

+++ controllers/contactController.js
@@ -55,6 +55,13 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not Found!");
   }
 
+  if (contact.user_id.toString() === req.user.id) {
+    res.status(403);
+    throw new Error(
+      "Users don't have permission to update the other user's Contacts."
+    );
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -74,6 +81,13 @@ const deleteContact = asyncHandler(async (req, res) => {
     console.log("Got in!");
     res.status(404);
     throw new Error("Contact not Found!");
+  }
+
+  if (contact.user_id.toString() === req.user.id) {
+    res.status(403);
+    throw new Error(
+      "Users don't have permission to delete the other user's Contacts."
+    );
   }
 
   await Contact.findByIdAndRemove(req.params.id);
